+++ src/lib/pages/execute/components/JsonExecute.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex, Button } from "@chakra-ui/react";
+import { Box, Flex } from "@chakra-ui/react";
 import type { Coin, StdFee } from "@cosmjs/stargate";
 import dynamic from "next/dynamic";
 import { useCallback, useEffect, useMemo, useState } from "react";
@@ -9,11 +9,10 @@ import {
   useFabricateFee,
   useExecuteContractTx,
   useCurrentChain,
-  useMobile,
-  useIsMac,
 } from "lib/app-provider";
 import { useAttachFunds } from "lib/app-provider/hooks/useAttachFunds";
 import { useSimulateFeeQuery } from "lib/app-provider/queries";
+import { SubmitButton } from "lib/components/button";
 import { CopyButton } from "lib/components/copy";
 import { ErrorMessageRender } from "lib/components/ErrorMessageRender";
 import { EstimatedFeeRender } from "lib/components/EstimatedFeeRender";
@@ -26,7 +25,6 @@ import {
 } from "lib/components/fund/data";
 import type { AttachFundsState } from "lib/components/fund/types";
 import { AttachFundsType } from "lib/components/fund/types";
-import { CustomIcon } from "lib/components/icon";
 import JsonInput from "lib/components/json/JsonInput";
 import { LoadingOverlay } from "lib/components/LoadingOverlay";
 import { useExecuteCmds } from "lib/hooks";
@@ -66,8 +64,6 @@ export const JsonExecute = ({
   // ------------------------------------------//
   // --------------DEPENDENCIES----------------//
   // ------------------------------------------//
-  const isMobile = useMobile();
-  const isMac = useIsMac();
   const { address } = useCurrentChain();
   const fabricateFee = useFabricateFee();
   const executeTx = useExecuteContractTx();
@@ -243,20 +239,6 @@ export const JsonExecute = ({
   ]);
 
   const isButtonDisabled = !enableExecute || !fee || isFetching;
-  useEffect(() => {
-    const keydownHandler = (e: KeyboardEvent) => {
-      // TODO: problem with safari if focusing in the textarea
-      const specialKey = isMac ? e.metaKey : e.ctrlKey;
-      if (!isButtonDisabled && specialKey && e.key === "Enter") {
-        proceed();
-      }
-    };
-    document.addEventListener("keydown", keydownHandler);
-    return () => {
-      document.removeEventListener("keydown", keydownHandler);
-    };
-  });
-
   return (
     <>
       {cmdsFetching && <LoadingOverlay />}
@@ -296,18 +278,12 @@ export const JsonExecute = ({
             Transaction Fee:{" "}
             <EstimatedFeeRender estimatedFee={fee} loading={isFetching} />
           </Flex>
-          <Button
-            variant="primary"
-            fontSize="14px"
-            p="6px 16px"
-            onClick={proceed}
-            isDisabled={isButtonDisabled}
-            leftIcon={<CustomIcon name="execute" />}
+          <SubmitButton
+            text="Execute"
             isLoading={processing}
-            sx={{ pointerEvents: processing && "none" }}
-          >
-            Execute {!isMobile && ` (${isMac ? "⌘" : "Ctrl"} + Enter)`}
-          </Button>
+            onSubmit={proceed}
+            isDisabled={isButtonDisabled}
+          />
         </Flex>
       </Flex>
     </>
