+++ src/lib/pages/query/components/JsonQuery.tsx
@@ -1,4 +1,4 @@
-import { ButtonGroup, Flex, Button, Spacer, Box, Text } from "@chakra-ui/react";
+import { ButtonGroup, Flex, Spacer, Box, Text } from "@chakra-ui/react";
 import { useQuery } from "@tanstack/react-query";
 import type { AxiosError } from "axios";
 import dynamic from "next/dynamic";
@@ -9,12 +9,10 @@ import {
   CELATONE_QUERY_KEYS,
   useBaseApiRoute,
   useCurrentChain,
-  useIsMac,
-  useMobile,
 } from "lib/app-provider";
+import { SubmitButton } from "lib/components/button";
 import { ContractCmdButton } from "lib/components/ContractCmdButton";
 import { CopyButton } from "lib/components/copy";
-import { CustomIcon } from "lib/components/icon";
 import JsonInput from "lib/components/json/JsonInput";
 import JsonReadOnly from "lib/components/json/JsonReadOnly";
 import { LoadingOverlay } from "lib/components/LoadingOverlay";
@@ -45,8 +43,6 @@ interface JsonQueryProps {
 
 export const JsonQuery = ({ contractAddress, initialMsg }: JsonQueryProps) => {
   const { track, trackAction } = useTrack();
-  const isMobile = useMobile();
-  const isMac = useIsMac();
   const { isFetching: cmdsFetching, queryCmds } = useQueryCmds(contractAddress);
   const lcdEndpoint = useBaseApiRoute("rest");
   const { addActivity } = useContractStore();
@@ -94,18 +90,6 @@ export const JsonQuery = ({ contractAddress, initialMsg }: JsonQueryProps) => {
   };
 
   const isButtonDisabled = jsonValidate(msg) !== null;
-  useEffect(() => {
-    const keydownHandler = (e: KeyboardEvent) => {
-      // TODO: problem with safari if focusing in the textarea
-      const specialKey = isMac ? e.metaKey : e.ctrlKey;
-      if (!isButtonDisabled && specialKey && e.key === "Enter") handleQuery();
-    };
-    document.addEventListener("keydown", keydownHandler);
-    return () => {
-      document.removeEventListener("keydown", keydownHandler);
-    };
-  });
-
   return (
     <>
       {cmdsFetching && <LoadingOverlay />}
@@ -162,18 +146,12 @@ export const JsonQuery = ({ contractAddress, initialMsg }: JsonQueryProps) => {
                 message={msg}
               />
             </Flex>
-            <Button
-              variant="primary"
-              fontSize="14px"
-              p="6px 16px"
-              size={{ base: "sm", md: "md" }}
-              onClick={handleQuery}
-              isDisabled={isButtonDisabled}
+            <SubmitButton
+              text="Query"
               isLoading={queryFetching || queryRefetching}
-              leftIcon={<CustomIcon name="query" />}
-            >
-              Query{!isMobile && ` (${isMac ? "⌘" : "Ctrl"} + Enter)`}
-            </Button>
+              onSubmit={handleQuery}
+              isDisabled={isButtonDisabled}
+            />
           </Flex>
         </Box>
         <Spacer
